geocoding/location-service/pyproject.toml
[tool.poetry]
name = "location-service"
version = "1.0.0"
description = "Geocoding for global.health"
authors = ["Global.health: a Data Science Initiative <ghdsi.info@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
attrs = "20.3.0"
boto3 = "1.17.57"
botocore = "1.20.57"
CacheControl = "0.12.6"
certifi = "2020.12.5"
chardet = "4.0.0"
click = "7.1.2"
Flask = "1.1.2"
Flask-Inputs = "0.3.0"
future = "0.18.2"
idna = "2.10"
inflection = "0.2.0"
iniconfig = "1.1.1"
iso3166 = "1.0.1"
itsdangerous = "1.1.0"
Jinja2 = "2.11.3"
jmespath = "0.10.0"
json_schema = "0.3"
lru-dict = "1.1.7"
mapbox = "0.18.0"
MarkupSafe = "1.1.1"
mock = "4.0.3"
msgpack = "1.0.2"
packaging = "20.9"
pluggy = "0.13.1"
polyline = "1.4.0"
py = "1.10.0"
pymongo = "3.11.3"
pyparsing = "2.4.7"
python-dateutil = "2.8.1"
requests = "2.25.1"
s3transfer = "0.4.2"
Schemer = "0.2.11"
six = "1.15.0"
toml = "0.10.2"
uritemplate = "3.0.1"
urllib3 = "1.26.4"
Werkzeug = "1.0.1"
WTForms = "2.3.3"
ratelimiter = "^1.2.0"

[tool.poetry.dev-dependencies]
pymongo-inmemory = "0.2.0"
pytest = "6.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
